requirements.txt
dlib==19.19.0
Flask==0.12.2
Jinja2==2.11.2
numpy==1.18.2
Pillow==7.2.0
opencv-python==4.2.0.32
Werkzeug==0.16.1
astroid==2.3.3
Click==7.0
colorama==0.4.3
fastapi==0.47.1
gunicorn==20.0.4
h11==0.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pydantic==1.4
pylint==2.4.4
six==1.14.0
starlette==0.12.9
uvicorn==0.11.2
websockets==8.1
wrapt==1.11.2
